gacha_oddsテーブルに関するメソッドのみに修正
